--- src/main/java/org/seedstack/coffig/data/mutable/MutableTreeNodeFactory.java
@@ -1,23 +1,0 @@
-/**
- * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package org.seedstack.coffig.data.mutable;
-
-import org.seedstack.coffig.data.TreeNode;
-
-class MutableTreeNodeFactory {
-
-    static MutableTreeNode createFromPrefix(String prefix) {
-        MutableTreeNode treeNode;
-        if (new Prefix(prefix).isArray) {
-            treeNode = new MutableArrayNode();
-        } else {
-            treeNode = new MutableMapNode();
-        }
-        return treeNode;
-    }
-}
